+++ scripts/download-wda-sim.mjs
@@ -2,7 +2,7 @@ import {fs, logger, zip, net, node} from 'appium/support.js';
 import _ from 'lodash';
 import os from 'os';
 import path from 'path';
-import { parseArgValue } from './utils.mjs';
+import { parseArgValue } from './utils.js';
 
 const log = logger.getLogger('download-wda-sim');
 const wdaUrl = (version, zipFileName) =>
